Do not ignore infra command from config files
It looks like this was previously removed because the default
hard-coded `/pause` so we would never take into account the image
config. I've removed the default in c/common and re-added support
to check config files.

While we're at it, fix ENTRYPOINT support - we should not be
setting this if we got ENTRYPOINT from the image.

Fixed https://bugzilla.redhat.com/show_bug.cgi?id=1853455

Signed-off-by: Matthew Heon <mheon@redhat.com>