Do not pull rootless CNI infra image
Instead, we want to advise users to manually build the image. We
cannot distribute the existing image for RHEL 8.3.1, and the
feature will be tech preview, so this degraded user experience
will have to be sufficient until we can get a better solution in
place.

Ref: https://issues.redhat.com/browse/RUN-1127

Please note that this is a RHEL only change and should not be
included in non-RHEL branches.

Signed-off-by: Matthew Heon <mheon@redhat.com>