play kube: fix args/command handling
when neither yaml.Args nor yaml.Command are specified, we should use the entrypoint and cmd from the image.

update the tests to cover this and another case (both args and command are specified).
use the registry image instead of redis, as it has both an entrypoint and command specified.
update the documentation around this handling to hopefully prevent regressions and confusion.

Signed-off-by: Peter Hunt <pehunt@redhat.com>