Use errors.Is for comparing ErrOCIRuntime errors
This simplifies checking for ErrOCIRuntime type of errors by using go's
built-in `errors.Is` function.

In order to make this practical, some errors now wrap `ErrOCIRuntime`.
Namely: `ErrOCIRuntimePermissionDenied` and `ErrOCIRuntimeNotFound`.

Given that some errors might now wrap `ErrOCIRuntime`, the error
formatting in the podman command's output was adjusted to take this into
account.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>