Add --if-not-exists option to podman network create
If `--if-not-exists` is set `podman network create` will not error
if the a network with the same given name already exists.

This option is only supported for local podman client. Supporting
the remote client would require changes in the api endpoint.

This option can be used as `ExecStartPre` command in a systemd
service file. see #8842

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>